[REM] wps: remove legacy routes and redirect
