+++ website_product_subscription/__openerp__.py
@@ -22,8 +22,9 @@
     "data": [
         "views/res_company.xml",
         "views/subscription_template_view.xml",
-        "views/product_subscription_template.xml",  # todo remove
+        "views/become_subscriber_menu.xml",
         "templates/components.xml",
+        "templates/subscribe_thanks.xml",
         "templates/subscribe_form.xml",
         "templates/subscribe_gift_form.xml",
         "templates/subscribe_generic_form.xml",
