+++ website_product_subscription/controllers/__init__.py
@@ -1,3 +1,2 @@
-from . import main
 from . import subscribe_form
 from . import subscribe
