--- website_product_subscription/tests/__init__.py
@@ -1,1 +1,0 @@
-from . import test_web_subscription
