+++ website_product_subscription/views/become_subscriber_menu.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+      Copyright 2019 Coop IT Easy
+      License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+-->
+<odoo>
+    <data>
+        <record id="menu_becomesubscriber" model="website.menu">
+            <field name="name">Become subscriber</field>
+            <field name="url">/new/subscription/generic</field>
+            <field name="parent_id" ref="website.main_menu"/>
+            <field name="sequence" type="int">50</field>
+        </record>
+    </data>
+</odoo>
